*: use obol's go-eth2-client improved fork (#3862)
We have opted for actively maintaining a fork of go-eth2-client. As we have added the missing functionality, now we are removing the duplicates in charon. Mind you, that some endpoints were already implemented by attestant's go-eth2-client, but at the time Charon needed them they weren't.

- Use go-eth2-client's block `getBlockAttestationsV2`, `getPeerCount`, `getEpochCommittees`, `getBlockV2`.
- Remove endpoints exposed to VC - `/proposer_config` and `/teku_proposer_config`. They were used in cluster locks <v1.7 (> 2 years ago).
- Update mockery files

category: feature
ticket: #3647

+++ core/fetcher/fetcher.go
@@ -324,7 +324,7 @@ func (f *Fetcher) fetchContributionData(ctx context.Context, slot uint64, defSet
 			return core.UnsignedDataSet{}, errors.New("invalid sync committee selection")
 		}
 
-		subcommIdx := uint64(selection.SubcommitteeIndex)
+		subcommIdx := selection.SubcommitteeIndex
 
 		// Check if the validator is an aggregator for the sync committee.
 		ok, err = eth2exp.IsSyncCommAggregator(ctx, f.eth2Cl, selection.SelectionProof)
